requirements.txt
requests
PyYAML
lxml
anticaptchaofficial